+++ SSD/file_handler_mock.h
@@ -9,6 +9,6 @@ public:
 	MOCK_METHOD(char*, readFile, (const string&), (override));
 	MOCK_METHOD(void, writeData, (const string&, const string&), (override));
 	MOCK_METHOD(void, rename, (const string&, const string&), (override, const));
-	MOCK_METHOD(bool, isFileExistByMatchLength, (const string&, const string&, int), (override));
+	MOCK_METHOD(bool, isFileExistByMatchLength, (const string&, const string&, size_t), (override));
 	MOCK_METHOD(vector<string>, getFileUsingPrefix, (const string&, const string&), (override));
 };
